broker/rusage: refactor rusage request handler
Problem: there are duplicate request handlers for cmb.rusage
and <module>.rusage

Combine broker cmb.rusage and <module>.rusage using an approach
similar to that used for ping.  Pull in the getrusage_json()
function since this is the only user.
